\name{samOptPars}
\alias{samOptPars}

\title{sam Parameter Optimization} 

\description{
  Function to optimize Sam parameter for maximim nuber of genes with a user defined FDR}
\usage{
samOptPars(opt, ths) 
}
\arguments{
  \item{opt}{an Sam optimization object}
  \item{ths}{a threshold for the FDR optimization}
}
\value{
  a numeric vector
}



\author{Claudio Isella, \email{claudio.isella@ircc.it}}
\keyword{MulCom}
